\name{Apo}
\alias{Apo}
\docType{data}
\title{apolipoprotein survey varaince component study}
\description{
Data from the apolipoprotein survey variance component study of Chapter 5 in
Design and Analysis of Experiments with R
}
\usage{data(Apo)}
\format{
  A data frame with 30 observations on the following 2 variables.
  \describe{
    \item{\code{lab}}{a factor with levels \code{A} \code{B}  \code{C} \code{D}}
	\item{\code{conc}}{a numeric vector}
  }
}

\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}
\examples{
data(Apo)
}
\keyword{datasets}
