\name{Bff}
\alias{Bff}
\docType{data}
\title{Confounded block fractional mouse growth experiment}
\description{
Data from the Confounded block fractional factorial mouse growth experiment in chapter 7 of Design and Analysis
of Experiments with R
}
\usage{data(Bff)}
\format{
  A data frame with 16 observations on the following 5 variables.
  \describe{
    \item{\code{Blocks}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5} \code{6} \code{7} \code{8}}
    \item{\code{A}}{a factor with levels \code{-1} \code{1}}
	\item{\code{B}}{a factor with levels \code{-1} \code{1}}
	\item{\code{C}}{a factor with levels \code{-1} \code{1}}
	\item{\code{D}}{a factor with levels \code{-1} \code{1}}
	\item{\code{E}}{a factor with levels \code{-1} \code{1}}
	\item{\code{F}}{a factor with levels \code{-1} \code{1}}	
	\item{\code{G}}{a factor with levels \code{-1} \code{1}}
	\item{\code{H}}{a factor with levels \code{-1} \code{1}}	
    \item{\code{weight}}{a numeric vector containing the response}
  }
}
\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}
\examples{
data(Bff)
}
\keyword{datasets}
