\name{blood}
\alias{blood}
\docType{data}
\title{ Variance component study of calcium in blood serum}
\description{
Data from the Variance component study of calcium in blood serum in chapter 5 of Design and Analysis
of Experiments with R
}
\usage{data(blood)}
\format{
  A data frame with 27 observations on the following 3 variables.
  \describe{
    \item{\code{sol}}{a factor with levels \code{1} \code{2} \code{3} \code{4} }
    \item{\code{lab}}{a factor with levels \code{A} \code{B} \code{C}}
    \item{\code{calcium}}{a numeric vector}
  }
}

\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}

\examples{
data(blood)
}
\keyword{datasets}
