\name{bread}
\alias{bread}
\docType{data}
\title{ Bread rise experiment data from Chapter 2}
\description{
Data from the bread rise experiment in chapter 2 of 
Design and Analysis of Experiments with R
}
\usage{data(bread)}
\format{
  A data frame with 12 observations on the following 3 variables.
  \describe{
    \item{\code{loaf}}{a numeric vector}
	\item{\code{time}}{a numeric vector}
    \item{\code{height}}{a numeric vector}
  }
}
\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}
\examples{
data(bread)
}
\keyword{datasets}
