\name{cakeb}
\alias{cakeb}
\docType{data}
\title{Split-Plot response surface for cake baking experiment}
\description{
Data from the Split-Plot response surface for cake baking experiment in chapter 10 of Design and Analysis
of Experiments with R 
}
\usage{data(cakeb)}
\format{
  A data frame with 11 observations on the following 6 variables.
  \describe{ 
 	\item{\code{Ovenrun}}{a factor with levels \code{1} \code{2} \code{3} \code{4}}
	\item{\code{x1}}{a numeric vector }
	\item{\code{x2}}{a numeric vector }
	\item{\code{y}}{a numeric vector }
	\item{\code{x1sq}}{a numeric vector }
	\item{\code{x2sq}}{a numeric vector }
  }
}
\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}
\examples{
data(cakeb)
}
\keyword{datasets}
