\name{castf}
\alias{castf}
\docType{data}
\title{cast fatigue experiment}
\description{
Data from the cast fatigue experiment in chapter 6 of Design and Analysis
of Experiments with R
}
\usage{data(castf)}
\format{
  A data frame with 12 observations on the following 12 variables.
  \describe{
    \item{\code{c8}}{a factor with levels \code{-1} \code{1} }
    \item{\code{c9}}{a factor with levels \code{-1} \code{1} }
    \item{\code{c10}}{a factor with levels \code{-1} \code{1} }
    \item{\code{c11}}{a factor with levels \code{-1} \code{1} }
	\item{\code{G}}{a factor with levels \code{-1} \code{1} }
	\item{\code{F}}{a factor with levels \code{-1} \code{1} }
	\item{\code{E}}{a factor with levels \code{-1} \code{1} }	
	\item{\code{D}}{a factor with levels \code{-1} \code{1} }
	\item{\code{C}}{a factor with levels \code{-1} \code{1} }
	\item{\code{B}}{a factor with levels \code{-1} \code{1} }
	\item{\code{A}}{a factor with levels \code{-1} \code{1} }			
    \item{\code{y}}{a numeric vector}
  }
}
\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}
\examples{
data(castf)
}
\keyword{datasets}
