\name{cement}
\alias{cement}
\docType{data}
\title{CCD design for cement workability experiment}
\description{
Data from the CCD design for cement workability experiment in chapter 10 of Design and Analysis
of Experiments with R 
}
\usage{data(cement)}
\format{
  A data frame with 20 observations on the following 4 variables.
  \describe{ 
 	\item{\code{Block}}{a factor with levels \code{1} \code{2}}
	\item{\code{x1}}{a numeric vector }
	\item{\code{x2}}{a numeric vector }
	\item{\code{x3}}{a numeric vector }
	\item{\code{y}}{a numeric vector }
  }
}
\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}
\examples{
data(cement)
}
\keyword{datasets}
