\name{chem}
\alias{chem}
\docType{data}
\title{Chemical process experiment data from Chapter 3}
\description{
Data from the Chemical process experiment  in chapter 3 of Design and Analysis
of Experiments with R
}
\usage{data(chem)}
\format{
  A data frame with 16 observations on the following 4 variables.
  \describe{
    \item{\code{A}}{a numeric vector containing the coded (-1,1) levels of factor A}
	\item{\code{B}}{a numeric vector containing the coded (-1,1) levels of factor B}
	\item{\code{C}}{a numeric vector containing the coded (-1,1) levels of factor C}
	\item{\code{D}}{a numeric vector containing the coded (-1,1) levels of factor D}
    \item{\code{y}}{a numeric vector containing the response}
  }
}
\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}
\examples{
data(chem)
}
\keyword{datasets}
