\name{COdata}
\alias{COdata}
\docType{data}
\title{ CO emmisions experiment data from Chapter 3}
\description{
Data from the CO emissions experiment in chapter 3 of Design and Analysis
of Experiments with R
}
\usage{data(COdata)}
\format{
  A data frame with 18 observations on the following 3 variables.
  \describe{
    \item{\code{Eth}}{a factor with levels \code{0.1} \code{0.2} \code{0.3}}
    \item{\code{Ratio}}{a factor with levels \code{14} \code{15} \code{16}}
    \item{\code{CO}}{a numeric vector}
  }
}

\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}
\examples{
data(COdata)
}
\keyword{datasets}
