\name{connector}
\alias{connector}
\docType{data}
\title{Table 12.21 Experiment with Elastometric Connector }
\description{
  Data from the Single Array Experiment with an Elastometric Connector in Chapter 12 of Design and Analysis
  of Experiments with R. The control and noise factors are in coded levels. 
}
\usage{data(connector)}
\format{
  A data frame with 32 observations on the following 8 variables. 
  \describe{ 
    \item{\code{A}}{a numeric vector }
    \item{\code{B}}{a numeric vector }
    \item{\code{C}}{a numeric vector }
    \item{\code{D}}{a numeric vector }
    \item{\code{E}}{a numeric vector }
    \item{\code{F}}{a numeric vector }
    \item{\code{G}}{a numeric vector }
    \item{\code{y}}{a numeric vector }
  }
}
\source{
  Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}
\examples{
  data(connector)
}
\keyword{datasets}
