\name{cont}
\alias{cont}
\docType{data}
\title{Control factor array and summary statistics for controller circuit design experiment}
\description{
Data from the control factor array and summary statistics for controller circuit design experiment in chapter 12 of Design and Analysis
of Experiments with R
}
\usage{data(cont)}
\format{
  A data frame with 18 observations on the following 6 variables.
  \describe{ 
	\item{\code{A}}{a numeric vector }
	\item{\code{B}}{a numeric vector }
	\item{\code{C}}{a numeric vector }
	\item{\code{D}}{a numeric vector }
	\item{\code{F}}{a numeric vector }
	\item{\code{lns2}}{a numeric vector }
  }
}
\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}
\examples{
data(cont)
}
\keyword{datasets}
