\name{cpipe}
\alias{cpipe}
\docType{data}
\title{Split-plot response surface for ceramic pipe experiment}
\description{
Data from the Split-plot response surface for ceramic pipe experiment in chapter 10 of Design and Analysis
of Experiments with R 
}
\usage{data(cpipe)}
\format{
  A data frame with 48 observations on the following 6 variables.
  \describe{ 
 	\item{\code{WP}}{a factor with levels \code{1} \code{2}  \code{3} \code{4} \code{5} \code{6}  \code{7} \code{8} \code{9} \code{10}  \code{11} \code{12}}
	\item{\code{A}}{a numeric vector }
	\item{\code{B}}{a numeric vector }
	\item{\code{P}}{a numeric vector }
	\item{\code{Q}}{a numeric vector }
	\item{\code{y}}{a numeric vector }
  }
}
\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}
\examples{
data(cpipe)
}
\keyword{datasets}
