\name{culture}
\alias{culture}
\docType{data}
\title{paecilomyces variotii culture experiment}
\description{
Data from the paecilomyces variotii culture experiment experiment in chapter 6 of Design and Analysis
of Experiments with R
}
\usage{data(culture)}
\format{
  A data frame with 16 observations on the following 9 variables.
  \describe{
    \item{\code{A}}{a factor with levels \code{-1} \code{1} }
    \item{\code{B}}{a factor with levels \code{-1} \code{1} }
	\item{\code{C}}{a factor with levels \code{-1} \code{1} }
	\item{\code{D}}{a factor with levels \code{-1} \code{1} }
	\item{\code{E}}{a factor with levels \code{-1} \code{1} }	
	\item{\code{F}}{a factor with levels \code{-1} \code{1} }
	\item{\code{G}}{a factor with levels \code{-1} \code{1} }
	\item{\code{H}}{a factor with levels \code{-1} \code{1} }		
    \item{\code{y1}}{a numeric vector}
  }
}
\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}
\examples{
data(culture)
}
\keyword{datasets}
