\name{D16F6}
\alias{D16F6}
\docType{data}
\title{Jones and Montgomery's 16 run alternative screening design for 6 Factors}
\description{
Design matrix for Jones and Montgomery's 16 run alternative screening design for 6 Factors
}
\usage{data(D16F6)}
\format{
  A data frame with 16 observations on the following 6 variables.
  \describe{
    \item{\code{A}}{a numeric vector containing the coded (-1,1) levels of factor A}
	\item{\code{B}}{a numeric vector containing the coded (-1,1) levels of factor B}
	\item{\code{C}}{a numeric vector containing the coded (-1,1) levels of factor C}
	\item{\code{D}}{a numeric vector containing the coded (-1,1) levels of factor D}
	\item{\code{E}}{a numeric vector containing the coded (-1,1) levels of factor E}	
	\item{\code{F}}{a numeric vector containing the coded (-1,1) levels of factor F}	
  }
}
\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}
\references{
Jones, B., Montgomery, D.C. (1998) "Alternatives to resolution IV screening designs in 16 runs", \emph{Int. J. Experimental Design and Process Optimisation}, 1, 285-295.
}
\examples{
data(D16F6)
}
\keyword{datasets}
