\name{daewr-package}
\alias{daewr-package}
\alias{daewr}
\docType{package}
\title{Data frames and functions for Design and Analysis of Experiments with R}
\description{
This package contains the data sets and functions from the book Design and Analysis of Experiments with R published by CRC in 2013.
}
\author{
John Lawson

Maintainer: John Lawson <lawsonjsl7net@gmail.com>

}
\references{
J. Lawson, Design and Analysis of Experiments with R, CRC 2013.
}
\keyword{ package }

