\name{dairy}
\alias{dairy}
\docType{data}
\title{Repeated measures study with dairy cow diets}
\description{
Data from the Repeated measures study with dairy cow diets in chapter 9 of Design and Analysis
of Experiments with R (compact format)
}
\usage{data(dairy)}
\format{
  A data frame with 120 observations on the following 5 variables.
  \describe{ 
 	\item{\code{Diet}}{a factor with levels \code{"Barley"} \code{"Mixed"} \code{"Lupins"}}
	\item{\code{pr1}}{a numeric vector }
	\item{\code{pr2}}{a numeric vector }
	\item{\code{pr3}}{a numeric vector }
	\item{\code{pr4}}{a numeric vector }
  }
}
\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}
\examples{
data(dairy)
}
\keyword{datasets}
