\name{drug}
\alias{drug}
\docType{data}
\title{Data from rat behavior experiment in Chapter 4}
\description{
Data from rat behavior experiment in Chapter 4 of Design and Analysis
of Experiments with R
}
\usage{data(drug)}
\format{
  A data frame with 50 observations on the following 3 variables.
  \describe{
    \item{\code{rat}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5} \code{6} \code{7} \code{8} \code{9} \code{10}}
    \item{\code{dose}}{a factor with levels \code{0.0} \code{0.5} \code{1.0} \code{1.5} \code{2.0}}
    \item{\code{rate}}{a numeric vector}
  }
}
\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}
\examples{
data(drug)
}
\keyword{datasets}
