\name{eptaxr}
\alias{eptaxr}
\docType{data}
\title{Single array and raw response for silicon layer growth experiment}
\description{
Data from the single array and raw response for silicon layer growth experiment in chapter 12 of Design and Analysis
of Experiments with R
}
\usage{data(eptaxr)}
\format{
  A data frame with 64 observations on the following 9 variables.
  \describe{ 
	\item{\code{A}}{a numeric vector }
	\item{\code{B}}{a numeric vector }
	\item{\code{C}}{a numeric vector }
	\item{\code{D}}{a numeric vector }
	\item{\code{E}}{a numeric vector }
	\item{\code{F}}{a numeric vector }
	\item{\code{G}}{a numeric vector }
	\item{\code{H}}{a numeric vector }	
	\item{\code{y}}{a numeric vector }
  }
}
\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}
\examples{
data(eptaxr)
}
\keyword{datasets}
