\name{eptaxyb}
\alias{eptaxyb}
\docType{data}
\title{Control array and mean response for silicon layer growth experiment}
\description{
Data from the control array and mean response for silicon layer growth experiment in chapter 12 of Design and Analysis
of Experiments with R
}
\usage{data(eptaxyb)}
\format{
  A data frame with 16 observations on the following 9 variables.
  \describe{ 
	\item{\code{A}}{a numeric vector }
	\item{\code{B}}{a numeric vector }
	\item{\code{C}}{a numeric vector }
	\item{\code{D}}{a numeric vector }
	\item{\code{E}}{a numeric vector }
	\item{\code{F}}{a numeric vector }
	\item{\code{G}}{a numeric vector }
	\item{\code{H}}{a numeric vector }	
	\item{\code{ybar}}{a numeric vector }
  }
}
\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}
\examples{
data(eptaxyb)
}
\keyword{datasets}
