\name{fullnormal}
\alias{fullnormal}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{This function makes a full normal plot of the elements of the vector called effects
}
\description{
This function makes a full normal plot of the elements of the vector called effects
}
\usage{
fullnormal(effects, labs, alpha = 0.05, refline = "TRUE")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{effects}{
input - vector of effects to be plotted  
}
  \item{labs}{
input - vector of labels of the effects to be plotted
}
  \item{alpha}{
input - alpha level for labeling of significant effects using Lenth statistic
}
  \item{refline}{
input - logical variable that indicates whether a reference line is added to the plot (default is "TRUE")
}
}
\author{
John Lawson
}


%% ~Make other sections like Warning with \section{Warning }{....} ~


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ hplot }

