\name{gagerr}
\alias{gagerr}
\docType{data}
\title{ Gauge R&R Study}
\description{
Data from the Gauge R&R Study in chapter 5 of Design and Analysis
of Experiments with R
}
\usage{data(gagerr)}
\format{
  A data frame with 60 observations on the following 3 variables.
  \describe{
    \item{\code{part}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5} \code{6} \code{7} \code{8} \code{9} \code{10} }
    \item{\code{oper}}{a factor with levels \code{1} \code{2} \code{3}}
    \item{\code{y}}{a numeric vector}
  }
}

\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}
\examples{
data(gagerr)
}
\keyword{datasets}
