\name{inject}
\alias{inject}
\docType{data}
\title{Single array for injection molding experiment}
\description{
Data from the single array for injection molding experiment in chapter 12 of Design and Analysis
of Experiments with R
}
\usage{data(inject)}
\format{
  A data frame with 20 observations on the following 8 variables.
  \describe{ 
	\item{\code{A}}{a numeric vector }
	\item{\code{B}}{a numeric vector }
	\item{\code{C}}{a numeric vector }
	\item{\code{D}}{a numeric vector }
	\item{\code{E}}{a numeric vector }
	\item{\code{F}}{a numeric vector }
	\item{\code{G}}{a numeric vector }
	\item{\code{shrinkage}}{a numeric vector }
  }
}

\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}

\examples{
data(inject)
}
\keyword{datasets}
