\name{interleave}
\alias{interleave}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Interleave vectors
}
\description{
interleaves two vectors
}
\usage{
interleave(v1,v2)
}
\arguments{
  \item{v1}{
input - first vector
}
  \item{v2}{
input - second vector
}
}
\value{
vector
}
