\name{MPV}
\alias{MPV}
\docType{data}
\title{mixture process variable experiment with mayonnaise}
\description{
Data from the mixture process variable experiment with mayonnaise in chapter 11 of Design and Analysis
of Experiments with R 
}
\usage{data(MPV)}
\format{
  A data frame with 35 observations on the following 4 variables.
  \describe{ 
	\item{\code{x1}}{a numeric vector }
	\item{\code{x2}}{a numeric vector }
	\item{\code{x3}}{a numeric vector }
	\item{\code{z1}}{a numeric vector }
	\item{\code{z2}}{a numeric vector }
	\item{\code{y}}{a numeric vector }
  }
}

\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}

\examples{
data(MPV)
}
\keyword{datasets}
