\name{MR16m10g3}
\alias{MR16m10g3}
\docType{data}
\title{Li and Nachtsheim's Model Robust Screening Design for 10 factors in 16 runs}
\description{
Design matrix for Li and Nachtsheim's Model Robust Screening Design for 10 factors in 16 runs
}
\usage{data(MR16m10g3)}
\format{
  A data frame with 16 observations on the following 6 variables.
  \describe{
    \item{\code{A}}{a numeric vector containing the coded (-1,1) levels of factor A}
	\item{\code{B}}{a numeric vector containing the coded (-1,1) levels of factor B}
	\item{\code{C}}{a numeric vector containing the coded (-1,1) levels of factor C}
	\item{\code{D}}{a numeric vector containing the coded (-1,1) levels of factor D}
	\item{\code{E}}{a numeric vector containing the coded (-1,1) levels of factor E}
	\item{\code{F}}{a numeric vector containing the coded (-1,1) levels of factor F}
	\item{\code{G}}{a numeric vector containing the coded (-1,1) levels of factor G}
	\item{\code{H}}{a numeric vector containing the coded (-1,1) levels of factor H}
	\item{\code{J}}{a numeric vector containing the coded (-1,1) levels of factor J}	
    \item{\code{K}}{a numeric vector containing the coded (-1,1) levels of factor K}	
  }
}
\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}
\references{
Li W. and Nachtsheim C. (2000) "Model Robust Factorial Designs", \emph{Technometrics}, 42, 345-352.
}
\examples{
data(MR16m10g3)
}
\keyword{datasets}
