\name{MR8m6g1}
\alias{MR8m6g1}
\docType{data}
\title{Li and Nachtsheim's Model Robust Screening Design for 6 factors in 8 runs}
\description{
Design matrix for Li and Nachtsheim's Model Robust Screening Design for 6 factors in 8 runs
}
\usage{data(MR8m6g1)}
\format{
  A data frame with 16 observations on the following 6 variables.
  \describe{
    \item{\code{A}}{a numeric vector containing the coded (-1,1) levels of factor A}
	\item{\code{B}}{a numeric vector containing the coded (-1,1) levels of factor B}
	\item{\code{C}}{a numeric vector containing the coded (-1,1) levels of factor C}
	\item{\code{D}}{a numeric vector containing the coded (-1,1) levels of factor D}
	\item{\code{E}}{a numeric vector containing the coded (-1,1) levels of factor E}
	\item{\code{F}}{a numeric vector containing the coded (-1,1) levels of factor F}	
  }
}
\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}
\references{
Li W. and Nachtsheim C. (2000) "Model Robust Factorial Designs", \emph{Technometrics}, 42, 345-352.
}
\examples{
data(MR8m6g1)
}
\keyword{datasets}
