\name{Naph}
\alias{Naph}
\docType{data}
\title{Yields of naphthalene black}
\description{
Data from the Yields of naphthalene black of Chapter 5 in
Design and Analysis of Experiments with R
}
\usage{data(Naph)}
\format{
  A data frame with 30 observations on the following 2 variables.
  \describe{
    \item{\code{sample}}{a factor with levels \code{1} \code{2}  \code{3} \code{4} \code{5} \code{6}}
	\item{\code{yield}}{a numeric vector}
  }
}

\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}

\examples{
data(Naph)
}
\keyword{datasets}
