\name{pastry}
\alias{pastry}
\docType{data}
\title{Blocked response surface design for pastry dough experiment}
\description{
Data from the Blocked response surface design for pastry dough experiment in chapter 10 of Design and Analysis
of Experiments with R 
}
\usage{data(pastry)}
\format{
  A data frame with 28 observations on the following 5 variables.
  \describe{ 
 	\item{\code{Block}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5} \code{6} \code{7}}
	\item{\code{x1}}{a numeric vector }
	\item{\code{x2}}{a numeric vector }
	\item{\code{x3}}{a numeric vector }
	\item{\code{y}}{a numeric vector }
  }
}

\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}

\examples{
data(pastry)
}
\keyword{datasets}
