\name{pest}
\alias{pest}
\docType{data}
\title{Pesticide formulation experiment}
\description{
Data from the Pesticide formulation experiment in chapter 11 of Design and Analysis
of Experiments with R 
}
\usage{data(pest)}
\format{
  A data frame with 13 observations on the following 4 variables.
  \describe{ 
	\item{\code{x1}}{a numeric vector }
	\item{\code{x2}}{a numeric vector }
	\item{\code{x3}}{a numeric vector }
	\item{\code{y}}{a numeric vector }
  }
}

\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}

\examples{
data(pest)
}
\keyword{datasets}
