\name{plasma}
\alias{plasma}
\docType{data}
\title{Unreplicated split-plot \eqn{2^5} experiment on plasma treatment of paper}
\description{
Data from the unreplicated split-plot \eqn{2^5} experiment on plasma treatment of paper in chapter 8 of Design and Analysis
of Experiments with R
}
\usage{data(plasma)}
\format{
  A data frame with 32 observations on the following 6 variables.
  \describe{
    \item{\code{A}}{a factor with levels \code{-1} \code{1}}
	\item{\code{B}}{a factor with levels \code{-1} \code{1}}
	\item{\code{C}}{a factor with levels \code{-1} \code{1}}
	\item{\code{D}}{a factor with levels \code{-1} \code{1}}
	\item{\code{E}}{a factor with levels \code{-1} \code{1}}
    \item{\code{y}}{a numeric vector containing the response}
  }
}

\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}

\examples{
data(plasma)
}
\keyword{datasets}
