\name{polvdat}
\alias{polvdat}
\docType{data}
\title{Polvoron mixture experiment}
\description{
Data from the Polvoron mixture experiment in chapter 11 of Design and Analysis
of Experiments with R 
}
\usage{data(polvdat)}
\format{
  A data frame with 12 observations on the following 4 variables.
  \describe{ 
	\item{\code{x1}}{a numeric vector }
	\item{\code{x2}}{a numeric vector }
	\item{\code{x3}}{a numeric vector }
	\item{\code{y}}{a numeric vector }
  }
}

\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}

\examples{
data(polvdat)
}
\keyword{datasets}
