\name{polymer}
\alias{polymer}
\docType{data}
\title{polymerization strength variability study}
\description{
Data from the polymerization strength variability study in chapter 5 of Design and Analysis
of Experiments with R
}
\usage{data(polymer)}
\format{
  A data frame with 120 observations on the following 5 variables.
  \describe{
    \item{\code{lot}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5} \code{6} \code{7} \code{8} \code{9} \code{10} \code{11} \code{12} \code{13} \code{14} \code{15} \code{16} \code{17} \code{18} \code{19} \code{20} \code{21} \code{22} \code{23} \code{24} \code{25} \code{26} \code{27} \code{28} \code{29} \code{30}}
    \item{\code{box}}{a factor with levels \code{1} \code{2} }
	\item{\code{prep}}{a factor with levels \code{1} \code{2}}
	\item{\code{test}}{a factor with levels \code{1} \code{2} }
    \item{\code{strength}}{a numeric vector}
  }
}

\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}

\examples{
data(polymer)
}
\keyword{datasets}
