\name{prodstd}
\alias{prodstd}
\docType{data}
\title{Complete control factor array and noise factor array for connector experiment}
\description{
Data from the complete control factor array and noise factor array for connector experiment in chapter 12 of Design and Analysis
of Experiments with R
}
\usage{data(prodstd)}
\format{
  A data frame with 16 observations on the following 16 variables.
  \describe{ 
	\item{\code{A}}{a numeric vector }
	\item{\code{B}}{a numeric vector }
	\item{\code{C}}{a numeric vector }
	\item{\code{D}}{a numeric vector }
	\item{\code{E}}{a numeric vector }
	\item{\code{F}}{a numeric vector }
	\item{\code{G}}{a numeric vector }
	\item{\code{Pof}}{a numeric vector }
  }
}

\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}

\examples{
data(prodstd)
}
\keyword{datasets}
