\name{qsar}
\alias{qsar}
\docType{data}
\title{Library of substituted hydroxyphenylurea compounds}
\description{
Data from the Library of substituted hydroxyphenylurea compounds in chapter 10 of Design and Analysis
of Experiments with R (compact format)
}
\usage{data(qsar)}
\format{
  A data frame with 36 observations on the following 4 variables.
  \describe{ 
	\item{\code{Compound}}{a numeric vector }
	\item{\code{HE}}{a numeric vector }
	\item{\code{DMz}}{a numeric vector }
	\item{\code{S0K}}{a numeric vector }
  }
}

\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}

\examples{
data(qsar)
}
\keyword{datasets}
