\name{Rations}
\alias{Rations}
\docType{data}
\title{Cattle rations design experiment data from Table 10.16}
\description{
  Data from the cattle rations design experiment in chapter 10 of Design and Analysis
  of Experiments with R
}
\usage{data(Rations)}
\format{
  A data frame with 45 observations on the following 4 variables.
  \describe{
    \item{\code{Block}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5} \code{6} \code{7} \code{8}}
    \item{\code{x1}}{a numeric vector}
    \item{\code{x2}}{a numeric vector}
    \item{\code{ADG}}{a numeric vector}
  }
}

\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}

\examples{
data(Rations)
}
\keyword{datasets}
