\name{rcb}
\alias{rcb}
\docType{data}
\title{generalized RCB golf driving experiment}
\description{
Data from the generalized RCB golf driving experiment in chapter 4 of Design and Analysis
of Experiments with R
}
\usage{data(rcb)}
\format{
  A data frame with 135 observations on the following 3 variables.
  \describe{ 
	\item{\code{id}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5} \code{6} \code{7} \code{8} \code{9}}
    \item{\code{teehgt}}{a factor with levels \code{1} \code{2}  \code{3}}
	\item{\code{cdistance}}{a numeric vector }
  }
}

\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}

\examples{
data(rcb)
}
\keyword{datasets}
