\name{Smotor}
\alias{Smotor}
\docType{data}
\title{Single array for starting motor experiment}
\description{
Data from the single array for starting motor experiment in chapter 12 of Design and Analysis
of Experiments with R
}
\usage{data(Smotor)}
\format{
  A data frame with 18 observations on the following 6 variables.
  \describe{ 
  	\item{\code{A}}{a factor with levels \code{1} \code{2} }
  	\item{\code{B}}{a factor with levels \code{1} \code{2} \code{3}}
  	\item{\code{C}}{a factor with levels \code{1} \code{2} \code{3}}
  	\item{\code{D}}{a factor with levels \code{1} \code{2} \code{3}}
  	\item{\code{E}}{a factor with levels \code{1} \code{2} }
	\item{\code{torque}}{a numeric vector }
  }
}

\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}

\examples{
data(Smotor)
}
\keyword{datasets}
