\name{soup}
\alias{soup}
\docType{data}
\title{dry mix soup experiment}
\description{
Data from the dry mix soup experiment in chapter 6 of Design and Analysis
of Experiments with R
}
\usage{data(soup)}
\format{
  A data frame with 16 observations on the following 6 variables.
  \describe{
    \item{\code{A}}{a factor with levels \code{-1} \code{1} }
    \item{\code{B}}{a factor with levels \code{-1} \code{1} }
	\item{\code{C}}{a factor with levels \code{-1} \code{1} }
	\item{\code{D}}{a factor with levels \code{-1} \code{1} }
	\item{\code{E}}{a factor with levels \code{-1} \code{1} }	
    \item{\code{y}}{a numeric vector}
  }
}

\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}

\examples{
data(soup)
}
\keyword{datasets}
