\name{soupmx}
\alias{soupmx}
\docType{data}
\title{dry soup mix variance component study}
\description{
Data from the dry soup mix variance component study of Chapter 5 in
Design and Analysis of Experiments with R
}
\usage{data(soupmx)}
\format{
  A data frame with 12 observations on the following 2 variables.
  \describe{
    \item{\code{batch}}{a factor with levels \code{1} \code{2}  \code{3} \code{4}}
	\item{\code{weight}}{a numeric vector}
  }
}

\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}

\examples{
data(soupmx)
}
\keyword{datasets}
