\name{SPMPV}
\alias{SPMPV}
\docType{data}
\title{Split-plot mixture process variable experiment with vinyl}
\description{
Data from the Split-plot mixture process variable experiment with vinyl in chapter 10 of Design and Analysis
of Experiments with R
}
\usage{data(SPMPV)}
\format{
  A data frame with 28 observations on the following 7 variables.
  \describe{ 
 	\item{\code{wp}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5} \code{6} \code{7}}
	\item{\code{z1}}{a numeric vector }
	\item{\code{z2}}{a numeric vector }
	\item{\code{x1}}{a numeric vector }
	\item{\code{x2}}{a numeric vector }
	\item{\code{x3}}{a numeric vector }
	\item{\code{y}}{a numeric vector }
  }
}

\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}

\examples{
data(SPMPV)
}
\keyword{datasets}
