\name{strungtile}
\alias{strungtile}
\docType{data}
\title{Strung out control factor array and raw response data for Ina tile experiment}
\description{
Data from the strung out control factor array and raw response data  for Ina tile experiment in chapter 12 of Design and Analysis
of Experiments with R
}
\usage{data(strungtile)}
\format{
  A data frame with 16 observations on the following 16 variables.
  \describe{ 
	\item{\code{A}}{a numeric vector }
	\item{\code{B}}{a numeric vector }
	\item{\code{C}}{a numeric vector }
	\item{\code{D}}{a numeric vector }
	\item{\code{E}}{a numeric vector }
	\item{\code{F}}{a numeric vector }
	\item{\code{G}}{a numeric vector }
	\item{\code{H}}{a numeric vector }
	\item{\code{AH}}{a numeric vector }
	\item{\code{BH}}{a numeric vector }	
	\item{\code{CH}}{a numeric vector }
	\item{\code{DH}}{a numeric vector }	
	\item{\code{EH}}{a numeric vector }
	\item{\code{FH}}{a numeric vector }	
	\item{\code{GH}}{a numeric vector }		
	\item{\code{y}}{a numeric vector }
  }
}

\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}

\examples{
data(strungtile)
}
\keyword{datasets}
