\name{sugarbeet}
\alias{sugarbeet}
\docType{data}
\title{Sugarbeet data from Chapter 2}
\description{
Sugarbeet data from chapter 2 of Design and Analysis of Experiments
with R
}
\usage{data(sugarbeet)}
\format{
  A data frame with 18 observations on the following 2 variables.
  \describe{
    \item{\code{treat}}{a factor with levels \code{A} \code{B} \code{C} \code{D}}
    \item{\code{yield}}{a numeric vector}
  }
}

\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}

\examples{
data(sugarbeet)
}
\keyword{datasets}
