\name{teach}
\alias{teach}
\docType{data}
\title{ Teaching experiment data from Chapter 2}
\description{
Data from the teaching experiment in chapter 2 of Design and Analysis
of Experiments with R
}
\usage{data(teach)}
\format{
  A data frame with 30 observations on the following 4 variables.
  \describe{
    \item{\code{class}}{a numeric vector}
    \item{\code{method}}{a factor with levels \code{1} \code{2} \code{3}}
    \item{\code{score}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5}}
    \item{\code{count}}{a numeric vector}
  }
}

\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}

\examples{
data(teach)
}
\keyword{datasets}
