\name{Tet}
\alias{Tet}
\docType{data}
\title{Tetracycline concentration in plasma}
\description{
Data from the Tetracycline concentration in plasma study in chapter 10 of Design and Analysis
of Experiments with R (compact format)
}
\usage{data(Tet)}
\format{
  A data frame with 9 observations on the following 2 variables.
  \describe{ 
	\item{\code{Time}}{a numeric vector }
	\item{\code{Conc}}{a numeric vector }
  }
}

\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}

\examples{
data(Tet)
}
\keyword{datasets}
