\name{tile}
\alias{tile}
\docType{data}
\title{Control factor array and summary statistics for Ina tile experiment}
\description{
Data from the control factor array and summary statistics for Ina tile experiment in chapter 12 of Design and Analysis
of Experiments with R
}
\usage{data(tile)}
\format{
  A data frame with 8 observations on the following 11 variables.
  \describe{ 
	\item{\code{A}}{a numeric vector }
	\item{\code{B}}{a numeric vector }
	\item{\code{C}}{a numeric vector }
	\item{\code{D}}{a numeric vector }
	\item{\code{E}}{a numeric vector }
	\item{\code{F}}{a numeric vector }
	\item{\code{G}}{a numeric vector }
	\item{\code{y1}}{a numeric vector }
	\item{\code{y2}}{a numeric vector }
	\item{\code{ybar}}{a numeric vector }	
	\item{\code{lns2}}{a numeric vector }
  }
}

\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}

\examples{
data(tile)
}
\keyword{datasets}
