\name{Treb}
\alias{Treb}
\docType{data}
\title{Box-Behnken design for trebuchet experiment}
\description{
Data from the Box-Behnken design for trebuchet experiment in chapter 10 of Design and Analysis
of Experiments with R 
}
\usage{data(Treb)}
\format{
  A data frame with 15 observations on the following 4 variables.
  \describe{ 
	\item{\code{x1}}{a numeric vector }
	\item{\code{x2}}{a numeric vector }
	\item{\code{x3}}{a numeric vector }
	\item{\code{y}}{a numeric vector }
  }
}

\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}

\examples{
data(Treb)
}
\keyword{datasets}
