\name{vinyl}
\alias{vinyl}
\docType{data}
\title{Vinysl plasticizer formulations experiment data}
\description{
  Data from vinyl plasticiser formulation experiment in chapter 11 of Design and Analysis
  of Experiments with R
}
\usage{data(vinyl)}
\format{
  A data frame with 40 observations on the following 7 variables.
  \describe{
    \item{\code{WP}}{a numeric vector}
    \item{\code{x1}}{a numeric vector}
    \item{\code{x2}}{a numeric vector}
    \item{\code{x3}}{a numeric vector}
    \item{\code{z1}}{a numeric vector}
    \item{\code{z2}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}

\source{
  Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}

\examples{
  data(vinyl)
}
\keyword{datasets}
