\name{virus}
\alias{virus}
\docType{data}
\title{ Assay of Viral Contamination experiment data from Chapter 3}
\description{
Data from the Assay of Viral Contamination experiment in chapter 3 of Design and Analysis
of Experiments with R
}
\usage{data(virus)}
\format{
  A data frame with 18 observations on the following 3 variables.
  \describe{
    \item{\code{y}}{a numeric vector }
    \item{\code{Sample}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5} \code{6}}
    \item{\code{Dilution}}{a factor with levels \code{3} \code{4} \code{5}}
  }
}

\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}

\examples{
data(virus)
}
\keyword{datasets}
