\name{volt}
\alias{volt}
\docType{data}
\title{Volt meter experiment data from Chapter 3}
\description{
Data from the Volt meter experiment in chapter 3 of Design and Analysis
of Experiments with R
}
\usage{data(volt)}
\format{
  A data frame with 16 observations on the following 3 variables.
  \describe{
    \item{\code{y}}{a numeric vector }
    \item{\code{A}}{a factor containing the levels (22, 32) of factor A}
    \item{\code{B}}{a factor containing the levels (0.5, 5.0) of factor B}
    \item{\code{C}}{a factor containing the levels (0.5, 5.0) of factor C}
  }
}

\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}

\examples{
data(volt)
}
\keyword{datasets}
