\name{web}
\alias{web}
\docType{data}
\title{ Web page design experiment data from Chapter 3}
\description{
Data from the web page design experiment in chapter 3 of Design and Analysis
of Experiments with R
}
\usage{data(web)}
\format{
  A data frame with 36 observations on the following 6 variables.
  \describe{
    \item{\code{A}}{a factor with levels \code{1} \code{2}}
	\item{\code{B}}{a factor with levels \code{1} \code{2}}
	\item{\code{C}}{a factor with levels \code{1} \code{2}}
	\item{\code{D}}{a factor with levels \code{1} \code{2}}
    \item{\code{visitors}}{a numeric vector}
    \item{\code{signup}}{a numeric vector}
  }
}

\source{
Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}

\examples{
data(web)
}
\keyword{datasets}
