\name{WeldS}
\alias{WeldS}
\docType{data}
\title{Table 12.24 Experiment with Weld Tensile Strength }
\description{
  Data from the Single Array Experiment in Exercise 5 of Chapter 12 in Design and Analysis
  of Experiments with R. The factors are in coded levels. 
}
\usage{data(WeldS)}
\format{
  A data frame with 16 observations on the following 16 variables. 
  \describe{ 
    \item{\code{D}}{a numeric vector }
    \item{\code{H}}{a numeric vector }
    \item{\code{G}}{a numeric vector }
    \item{\code{A}}{a numeric vector }
    \item{\code{F}}{a numeric vector }
    \item{\code{GH}}{a numeric vector }
    \item{\code{C}}{a numeric vector }
    \item{\code{B}}{a numeric vector }
    \item{\code{J}}{a numeric vector }
    \item{\code{E}}{a numeric vector }
    \item{\code{AC}}{a numeric vector }
    \item{\code{AH}}{a numeric vector }
    \item{\code{AG}}{a numeric vector }
    \item{\code{e1}}{a numeric vector }
    \item{\code{e2}}{a numeric vector }
    \item{\code{y}}{a numeric vector }
  }
}
\source{
  Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}
\examples{
  data(WeldS)
}
\keyword{datasets}
