\name{Altscreen}
\alias{Altscreen}
\title{Alternate 16 run screening designs}
\description{ Recalls Jones and Montgomery's 16 run screening designs from data frames
}
\usage{
Altscreen(nfac)
}
\arguments{
  \item{nfac}{ input- an integer
}
}

\value{ design
}
\references{
Jones, B. and Montgomery, D. C.  (2010) "Alternatives to resolution IV screening designs in 16 runs", Int. J. Experimental Design and Process Optimization, Vol 1, No. 4, 2010.
}
\author{
John Lawson
}

\examples{
Altscreen(6)
     }
\keyword{ datagen }
